Prototype Zero
This one is focused on reading. I think when the translation from paper to screen happened we only focused on the superficial functions of reading not the underlying intentions and motivations.

When we put it like this we realize, an article to read needs also to be an article to write. An article needs to be part of something I am writing