Prototype Zero Complete
Prototype Zero is meant to be a direct translation of the article from a PDF reader into a web app. This is meant to demonstrate the article as fluid. Now that the web app affords fluidity and the article takes the shape of its container we can start visualizing what more can be done to this.

Only thing missing now are images