Prototype One - in progress
Vertical scrolling - continuity
Horizontal switch - change section

On hover, a paragraph is foregrounded with the text being bold and the background growing darker. This is a point where I have lots of decisions. Do we do only the background darkening? And if so, how dark? Do we do only the font bold? If so, how bold? If the font is too bold, it occupies too much space and becomes

It is very important for the visual hierarchy to remain intact. In this case you will see that the paper's title stays visible at all times, as does the section's title. There is some dead space that shows demarcates between the elements.
For this, the whole screen must **NOT** be scrollable. Instead, only the material should be scrollable.

Another decision Im curious about is whether it makes sense to keep the arrow buttons on the top, near the header or at the middle of the verticle-center of the page?

These micro design decisions are going to change the entire experience.

Another question is, does it suffice to have the switch between sections one by one, or do we need the jump-to-section thingy too? How important is that to the reading process? How does that change reading?

Overall, all of these are useful controls. More control is good for the reader.