Prototype One: change Abstract viewing from hover to click
The idea is for this to be a conditional viewing. On hover, considering it is such a large width, it gets somewhat frustrating to have it pop up in my face repeatedly. Opening it is by click it, and it closes automatically. To prompt opening, the cursor changes form to a pointer, and the background changes color. This sends the message to the user to click.