One: Variable Margins
Another meta-interaction that allows the reader to change the article's structure. I often find it hard to read it when the text is too spread out. I realized this with obsidian when I saw them give me a lot of dead space on the side. That changed the game for me. Do people prefer narrower reading spaces or wider?