Prototype One: removing blunt hover
When first introduced the blunt hover is too much of a whiplash and it hurts the contrast I'm trying to draw. This way the transition from Prototype Zero and One is more pronounced, it highlights the restructure and the paragraph numbers.

But for paragraph numbers to hit, we still need to go back on to the second prototype. The biggest difference is felt when you have a) Hover, b) Paragraph numbers, c) writing alongside reading d) double click to reference.

Its the economy it establishes on top of the paper that is just... its like from walking to getting on a horse, I dont need to force myself to connect with my leg muscles and run faster, I need to control my horse to go faster, i need to floor the gas to go hard and go in.