Prototype Zero: Feedback
Feedback from other prototypes, I wanted to explore how this version could grow. Like I said last night, there *is* a charm to the simple long-scroll but its not as feature rich as the others. So what if we enhance this too?

I have added the margins, hover, bold, font family, and paragraph numbers. How does this change my reading in this version?