Prototype Two: version 2.2 -- Chat Notes
Introduces a chat like interface with notes on the side. Also introduces a few meta interactions: Highlight to auto-copy, double click in the chat window to paste. All markdown formatting works in the notes on the side and on top of that, you can download your notes in markdown format.

Centralizes the focus on the writing and what one derives from the reading, rather than the article itsel.f